requirements.txt
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
constantly==23.10.4
coverage==7.6.7
cryptography==44.0.0
daphne==4.1.2
dill==0.3.9
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.5.0
django-eventstream==5.3.1
django-grip==3.5.1
django-model-utils==5.0.0
django-timescaledb==0.2.13
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.16.1
gripcontrol==4.2.0
hyperlink==21.0.0
identify==2.6.2
idna==3.10
incremental==24.7.2
isort==5.13.2
MarkupSafe==3.0.2
mccabe==0.7.0
nodeenv==1.9.1
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
platformdirs==4.3.6
pre_commit==4.0.1
psycopg==3.2.3
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.1
tomlkit==0.13.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
Werkzeug==3.1.3
zope.interface==7.2
